Bug 2021342 - Load encryption key from console earlier
Session restore of about:logins could expose a race condition where
trying to access the primary password that should be filled from the
console's sent primary secret is done but the code loading the secret
from the console has not been ran yet.

Moving to browser-before-ui-startup startup phase.
